+++ scrooge-generator-tests/src/test/resources/gold_file_output_scala/com/twitter/scrooge/test/gold/thriftscala/AnotherException.scala
@@ -280,6 +280,18 @@ class AnotherException(
   }
 
   /**
+   * If the specified fields are optional, they are set to None.  Otherwise, if the fields are
+   * known, they are reverted to their default values; if the fields are unknown, they are removed
+   * from the passthroughFields map, if present.
+   */
+  def unsetFields(_fieldIds: Set[Short]): AnotherException = {
+    new AnotherException(
+      if (_fieldIds(1)) 0 else this.errorCode,
+      _passthroughFields -- _fieldIds
+    )
+  }
+
+  /**
    * If the specified field is optional, it is set to None.  Otherwise, if the field is
    * known, it is reverted to its default value; if the field is unknown, it is removed
    * from the passthroughFields map, if present.